+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 3.3.2
 ----------------
 - Fix SQL error in students data in Transcripts.php
+- Fatal error when no calendars setup yet in Calendar.php
 
 Changes in 3.3.1
 ----------------